Add time on/off controls and make zero pos vertical
A change of logic is incurred as of this commit: the motor's zero
position will be when the beam cover is vertical, i.e., covering the
beam. From then on, the beam control counts the number of seconds
configured by the user for the OFF time. When the OFF time has passed,
the shutter is moved and the ON time starts. When the ON time has
passed, the shutter is moved back to the OFF position and the cycle
repeats.