Disable btnRotateCCW when motor zero pos. found
The motor controller cannot rotate negative (in our case, CCW) when it
is at zero position, so the button needs to be disabled, as it is on
startup, when the motor is assumed to be at its start-up zero position.
Note that the motor is not actually at _our_ definition of zero on
start-up, rather at some position defined by gravity and the other
forces "that be".