Make log file slightly more readable
Adding some spaces in between the semicolon delimiters makes the file
more readable when using a text file reader, e.g. Notepad.

What gets printed when the file is created or opened has also been
changed. A small log file header is printed to files, listing the
columns of data (useful if the file is imported into Excel). If a file
is being opened via the btnOpenLogfile, some new lines are printed
before the header to the file being opened, to help identify when a new
invocation of the tool was made.